[dawn][native] Use size_t to store string command lengths.
Fixed: 499159695
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/304695
Commit-Queue: Corentin Wallez <cwallez@chromium.org>

+++ src/dawn/native/Commands.cpp
@@ -435,9 +435,9 @@ void SkipCommand(CommandIterator* commands, Command type) {
     }
 }
 
-const char* AddNullTerminatedString(CommandAllocator* allocator, StringView s, uint32_t* length) {
+const char* AddNullTerminatedString(CommandAllocator* allocator, StringView s, size_t* length) {
     std::string_view view = s;
-    *length = static_cast<uint32_t>(view.length());
+    *length = view.length();
 
     // Include extra null-terminator character. The string_view may not be null-terminated. It also
     // may already have a null-terminator inside of it, in which case adding the null-terminator is
